GT-37: [bug] fixed recursive structs
